Move compiler-dependent code into a module
