+++ latestVersion.sbt
@@ -1,1 +1,5 @@
 latestVersion in ThisBuild := "0.6.2"
+latestVersionInSeries in ThisBuild := None
+unreleasedModules in ThisBuild := Set(
+  "refined-eval"
+)
