+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.6.3-SNAPSHOT"
+version in ThisBuild := "0.7.0-SNAPSHOT"
